+++ SrcLib/core/fwData/include/fwData/TriangularMesh.hpp
@@ -45,11 +45,11 @@ public :
     /**
      * Typedef using for pointer function
      */
-    typedef ::boost::function<void      ( ::fwData::TriangularMesh*, int, double, double, double ) > set1ValueFunc;
+    typedef ::boost::function<void      ( ::fwData::TriangularMesh*, int, float, float, float ) > set1ValueFunc;
     typedef ::boost::function<double*   ( ::fwData::TriangularMesh*, int)  > get1ValueFunc;
     typedef ::boost::function<void      ( ::fwData::TriangularMesh*, int, int, int, int ) > set1IndexFunc;
     typedef ::boost::function<const int ( ::fwData::TriangularMesh* ) > get1IndexFunc;
-    typedef ::boost::function<void      ( ::fwData::TriangularMesh*, int, double, double, double ) > set1NormalFunc;
+    typedef ::boost::function<void      ( ::fwData::TriangularMesh*, int, float, float, float ) > set1NormalFunc;
     typedef ::boost::function<const int ( ::fwData::TriangularMesh* ) > get1NormalFunc;
     typedef ::boost::function<const int ( ::fwData::TriangularMesh* ) > getNumPointsFunc;
     typedef ::boost::function<const int ( ::fwData::TriangularMesh* ) > getNumCellsFunc;
@@ -92,7 +92,7 @@ public :
      */
     FWDATA_API void clearCells();
 
-    FWDATA_API void setOneVectorInPointList(int _iIndex, double _fX, double _fY, double _fZ);
+    FWDATA_API void setOneVectorInPointList(int _iIndex, float _fX, float _fY, float _fZ);
 
     FWDATA_API void setOneIndexInIndexList(int _iIndex, int _p1, int _p2, int _p3);
 
@@ -100,13 +100,13 @@ public :
 
     FWDATA_API const int getOneIndexFromIndexList(void) const;
 
-    FWDATA_API void setOneVectorInNormalList(int x, double, double, double);
+    FWDATA_API void setOneVectorInNormalList(int x, float, float, float);
 
     FWDATA_API int const getOneIndexFromNormalList(void) const;
 
-    FWDATA_API int const getNumPoints(void) const ;
+    FWDATA_API size_t const getNumPoints(void) const ;
 
-    FWDATA_API int const getNumCells(void) const ;
+    FWDATA_API size_t const getNumCells(void) const ;
 
 protected :
